Validate create conference
Remove AdminController and add Conference stuff to its own controller.
Remove conference pages from admin folder
Rename routes
