+++ app/Http/Controllers/PresentationsController.php
@@ -36,18 +36,6 @@ class PresentationsController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function index($status = "Approved"){
-
-        /*
-        This fetches the courses with presentations and iterate through them
-        on the html page. It's used to divide the presentations by course.
-        */
-        /** HTML IS NOT WORKIN BECAUSE I PASS THE COURSE AND THE ITERATE THERE,7
-        MAKING THIS QUERY ALMOST INVALID!!!!!!*/
-        // $courses = Course::where('offered_this_semester',true)->
-        //     whereHas('presentations', function ($query) use ($status) {
-        //         $query->where('conference_id', '=', get_current_conference_id())->
-        //             where('status', '=', strtoupper($status[0]));
-        //     })->paginate(5);
         $status = strtoupper($status[0]);
         $presentations = Presentation::where(
             'conference_id', '=', get_current_conference_id())->
